refactor: remove "+" directory prefix for bounded contexts
