+++ src/client/app/about/index.ts
@@ -3,4 +3,3 @@
  */
 export * from './about.component';
 export * from './about.routes';
-
