+++ src/client/app/app.module.ts
@@ -6,8 +6,8 @@ import { HttpModule } from '@angular/http';
 import { AppComponent } from './app.component';
 import { routes } from './app.routes';
 
-import { AboutModule } from './+about/about.module';
-import { HomeModule } from './+home/home.module';
+import { AboutModule } from './about/about.module';
+import { HomeModule } from './home/home.module';
 import { SharedModule } from './shared/shared.module';
 
 @NgModule({
