+++ src/client/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 
-import { AboutRoutes } from './+about/index';
-import { HomeRoutes } from './+home/index';
+import { AboutRoutes } from './about/index';
+import { HomeRoutes } from './home/index';
 
 export const routes: Routes = [
   ...HomeRoutes,
