+++ src/client/app/home/home.component.ts
@@ -37,10 +37,10 @@ export class HomeComponent implements OnInit {
    */
   getNames() {
     this.nameListService.get()
-                     .subscribe(
-                       names => this.names = names,
-                       error =>  this.errorMessage = <any>error
-                       );
+		     .subscribe(
+		       names => this.names = names,
+		       error =>  this.errorMessage = <any>error
+		       );
   }
 
   /**
