+++ src/client/app/home/index.ts
@@ -3,4 +3,3 @@
  */
 export * from './home.component';
 export * from './home.routes';
-
